requirements.txt
cairocffi==1.0.2
CairoSVG==2.4.0
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
colorclass==2.2.0
cssselect2==0.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==1.11.22
django-classy-tags==0.9.0
django-cms==3.6.0
django-filer==1.5.0
django-formtools==2.1
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
django-sekizai==1.0.0
django-treebeard==4.3
djangocms-admin-style==1.2.9
djangocms-attributes-field==1.1.0
djangocms-column==1.9.0
djangocms-file==2.3.0
djangocms-googlemap==1.3.0
djangocms-installer==1.1.0
djangocms-link==2.4.0
djangocms-picture==2.0.8
djangocms-snippet==2.1.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.0.5
docopt==0.6.2
easy-thumbnails==2.6
html5lib==1.0.1
idna==2.8
packaging==19.0
Pillow==6.1.0
pip-upgrader==1.4.15
pycparser==2.19
pyparsing==2.4.0
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2019.1
reportlab==3.5.23
requests==2.22.0
six==1.12.0
South==1.0.2
sqlparse==0.3.0
terminaltables==3.1.0
tinycss2==1.0.2
tzlocal==1.5.1
Unidecode==1.0.23
urllib3==1.25.3
Wand==0.5.4
webencodings==0.5.1
wincertstore==0.2
xhtml2pdf==0.2.3
